Another task, another class.
